balance fix: new expertise, churn rate and brand power calculations

+++ Assets/Utils/Marketing/BrandingCampaign.cs
@@ -5,9 +5,9 @@ namespace Assets.Utils
 {
     public static partial class MarketingUtils
     {
-        public static void AddBrandPower(GameEntity company, int power)
+        public static void AddBrandPower(GameEntity company, float power)
         {
-            var brandPower = (int)Mathf.Clamp(company.branding.BrandPower + power, 0, 100);
+            var brandPower = Mathf.Clamp(company.branding.BrandPower + power, 0, 100);
 
             company.ReplaceBranding(brandPower);
         }
